refactor(mvvm, mvi 분리를 위한 기본작업): mvvm, mvi 분리를 위한 작업 시작
