+++ presentation/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
     
     <application>
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".common.ui.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.NewWeatherOpenApi">
             <intent-filter>
