+++ presentation/src/main/java/com/android/sj/presentation/common/ui/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.android.sj.presentation.ui
+package com.android.sj.presentation.common.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -10,7 +10,7 @@ import androidx.compose.ui.Modifier
 import com.android.sj.domain.managers.LocationDataManager
 import com.android.sj.presentation.ui.compose.InitScreen
 import com.android.sj.presentation.ui.compose.loading.SplashDialogScreen
-import com.android.sj.presentation.ui.theme.NewWeatherOpenApiTheme
+import com.android.sj.presentation.common.ui.theme.NewWeatherOpenApiTheme
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
