+++ presentation/src/main/java/com/android/sj/presentation/common/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.android.sj.presentation.ui.theme
+package com.android.sj.presentation.common.ui.theme
 
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Typography
