+++ presentation/src/main/java/com/android/sj/presentation/mvvm/viewmodels/NaverMapViewModel.kt
@@ -1,4 +1,4 @@
-package com.android.sj.presentation.viewmodels
+package com.android.sj.presentation.mvvm.viewmodels
 
 import android.content.Context
 import androidx.lifecycle.viewModelScope
