+++ presentation/src/main/java/com/android/sj/presentation/ui/compose/airQuality/AirQualityScreen.kt
@@ -11,8 +11,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
 import com.android.sj.presentation.R
 import com.android.sj.presentation.ui.compose.common.DotLineColumn
-import com.android.sj.presentation.ui.theme.Default_BackGround
-import com.android.sj.presentation.viewmodels.AirQualityViewModel
+import com.android.sj.presentation.common.ui.theme.Default_BackGround
+import com.android.sj.presentation.mvvm.viewmodels.AirQualityViewModel
 
 @Composable
 fun AirQualityScreen(
