+++ presentation/src/main/java/com/android/sj/presentation/ui/compose/airQuality/PredictionModelColumn.kt
@@ -22,7 +22,7 @@ import com.android.sj.presentation.R
 import com.android.sj.presentation.models.state.viewstate.AirQualityViewState
 import com.android.sj.presentation.ui.compose.common.UiStateHandler
 import com.android.sj.presentation.ui.previewParam.AirQualityPreviewParamProvider
-import com.android.sj.presentation.ui.theme.defaultTitleTextStyle
+import com.android.sj.presentation.common.ui.theme.defaultTitleTextStyle
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
 import kotlin.math.absoluteValue
