+++ presentation/src/main/java/com/android/sj/presentation/ui/compose/common/UiStateHandler.kt
@@ -15,8 +15,8 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import com.android.sj.presentation.R
 import com.android.sj.presentation.models.state.uistate.BaseUiState
-import com.android.sj.presentation.ui.theme.icon.EmptyImageVector
-import com.android.sj.presentation.ui.theme.icon.ErrorImageVector
+import com.android.sj.presentation.common.ui.theme.icon.EmptyImageVector
+import com.android.sj.presentation.common.ui.theme.icon.ErrorImageVector
 import com.android.sj.presentation.utils.dataPotalResultCode
 
 @Composable
