+++ presentation/src/main/java/com/android/sj/presentation/ui/compose/weather/TimeWeatherColumn.kt
@@ -34,7 +34,7 @@ import com.android.sj.presentation.models.state.viewstate.WeatherViewState
 import com.android.sj.presentation.models.uimodels.weather.TimeWeatherUIModel
 import com.android.sj.presentation.ui.compose.common.UiStateHandler
 import com.android.sj.presentation.ui.previewParam.CardDataPreviewParamProvider
-import com.android.sj.presentation.ui.theme.Color_eceff1
+import com.android.sj.presentation.common.ui.theme.Color_eceff1
 import com.android.sj.presentation.utils.sp
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
