+++ presentation/src/main/java/com/android/sj/presentation/ui/compose/weather/WeatherNowColumn.kt
@@ -25,7 +25,7 @@ import com.android.sj.presentation.R
 import com.android.sj.presentation.models.state.viewstate.WeatherViewState
 import com.android.sj.presentation.ui.compose.common.UiStateHandler
 import com.android.sj.presentation.ui.previewParam.WeatherViewStatePreviewParamProvider
-import com.android.sj.presentation.ui.theme.defaultTitleTextStyle
+import com.android.sj.presentation.common.ui.theme.defaultTitleTextStyle
 import com.android.sj.presentation.utils.sp
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
