+++ presentation/src/main/java/com/android/sj/presentation/ui/compose/weather/WeatherScreen.kt
@@ -13,8 +13,8 @@ import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.unit.dp
 import com.android.sj.presentation.R
 import com.android.sj.presentation.ui.compose.common.DotLineColumn
-import com.android.sj.presentation.ui.theme.Default_BackGround
-import com.android.sj.presentation.viewmodels.WeatherViewModel
+import com.android.sj.presentation.common.ui.theme.Default_BackGround
+import com.android.sj.presentation.mvvm.viewmodels.WeatherViewModel
 
 @Composable
 fun WeatherScreen(
